Entregas - Lucas Pereira/Projeto Final/requirements.txt
google-generativeai>=0.3.2
ipykernel>=6.29.0
python-dotenv
roboflow
opencv-python
ultralytics
supervision